feature: connector types
