+++ scaleph-ui-react2/src/pages/Project/Workspace/Artifact/DI/DiJobFlow/Dag/node/base-node.tsx
@@ -1,20 +1,45 @@
 import {Button, Descriptions, Popover, Space, Tag, Typography} from 'antd';
-import {InfoCircleOutlined, MenuOutlined} from '@ant-design/icons';
+import {HolderOutlined, InfoCircleOutlined, MenuOutlined} from '@ant-design/icons';
 import {useIntl} from "@umijs/max";
+import {useDnd} from "@antv/xflow";
 import './base-node.less';
 import {Props} from "@/typings";
+import styles from "@/pages/Project/Workspace/Artifact/DI/DiJobFlow/Dag/dnd/dnd.less";
+import React from "react";
+import {DAG_NODE} from "@/pages/Project/Workspace/Artifact/DI/DiJobFlow/Dag/shape";
 
 export const BaseNode: React.FC<Props<Record<string, any>>> = ({data}) => {
   const intl = useIntl();
+  const {startDrag} = useDnd();
+  let id = 0;
+
+  const handleMouseDown = (
+    e: React.MouseEvent<Element, MouseEvent>,
+  ) => {
+    id += 1;
+    startDrag(
+      {
+        id: id.toString(),
+        shape: DAG_NODE,
+        data: {
+          id: id.toString(),
+          label: data.title,
+          status: 'default',
+        },
+        ports: data.ports,
+      },
+      e,
+    );
+  };
 
   const nodeStyle = () => {
     switch (data.meta?.type) {
       case 'source':
-        return '#e6f4ff';
+        // return '#e6f4ff';
       case 'transform':
-        return '#e6fffb';
+        // return '#e6fffb';
       case 'sink':
-        return '#fff7e6';
+        // return '#fff7e6';
       default:
         return '';
     }
@@ -44,7 +69,7 @@ export const BaseNode: React.FC<Props<Record<string, any>>> = ({data}) => {
       }
       placement="rightTop"
     >
-      <div className="base-node" style={{backgroundColor: nodeStyle()}}>
+      <div className="base-node" style={{backgroundColor: nodeStyle()}} onMouseDown={(e) => handleMouseDown(e)}>
         <span className="icon">
           <MenuOutlined style={{color: '#3057e3', fontSize: '16px'}}/>
         </span>
@@ -53,6 +78,9 @@ export const BaseNode: React.FC<Props<Record<string, any>>> = ({data}) => {
             <Typography.Text ellipsis={true}>{data.title}</Typography.Text>
           </Space>
         </span>
+        <div className={styles.nodeDragHolder}>
+          <HolderOutlined/>
+        </div>
       </div>
     </Popover>
   );
